+++ engine/runtime/src/main/java/org/enso/interpreter/node/callable/resolver/HostMethodCallNode.java
@@ -98,13 +98,13 @@ public abstract class HostMethodCallNode extends Node {
    * used.
    *
    * @param self the method call target
-   * @param methodName the method name
+   * @param symbol symbol representing method to be resolved
    * @param library an instance of interop library to use for interacting with the target
    * @return a {@link PolyglotCallType} to use for this target and method
    */
   public static PolyglotCallType getPolyglotCallType(
-      Object self, String methodName, InteropLibrary library) {
-    return getPolyglotCallType(self, methodName, library, null);
+      Object self, UnresolvedSymbol symbol, InteropLibrary library) {
+    return getPolyglotCallType(self, symbol, library, null);
   }
 
   /**
@@ -112,14 +112,14 @@ public abstract class HostMethodCallNode extends Node {
    * used.
    *
    * @param self the method call target
-   * @param methodName the method name
+   * @param symbol symbol representing method to be resolved
    * @param library an instance of interop library to use for interacting with the target
    * @param methodResolverNode {@code null} or real instances of the node to resolve methods
    * @return a {@link PolyglotCallType} to use for this target and method
    */
   public static PolyglotCallType getPolyglotCallType(
       Object self,
-      String methodName,
+      UnresolvedSymbol symbol,
       InteropLibrary library,
       MethodResolverNode methodResolverNode) {
     if (library.isDate(self)) {
@@ -142,7 +142,6 @@ public abstract class HostMethodCallNode extends Node {
       if (methodResolverNode != null) {
         var ctx = Context.get(library);
         var arrayType = ctx.getBuiltins().array();
-        var symbol = UnresolvedSymbol.build(methodName, ctx.getBuiltins().getScope());
         var fn = methodResolverNode.execute(arrayType, symbol);
         if (fn != null) {
           return PolyglotCallType.CONVERT_TO_ARRAY;
@@ -152,6 +151,7 @@ public abstract class HostMethodCallNode extends Node {
       }
     }
 
+    String methodName = symbol.getName();
     if (library.isMemberInvocable(self, methodName)) {
       return PolyglotCallType.CALL_METHOD;
     } else if (library.isMemberReadable(self, methodName)) {